+++ test/condition/none-condition-test.js
@@ -5,7 +5,8 @@ var lewd = require('../../src/lewd'),
     errorMessages = require('../../src/messages'),
     helper = require('./../helper');
 
-var refuteValues = helper.refuteValues,
+var assert = buster.referee.assert,
+    refuteValues = helper.refuteValues,
     acceptValues = helper.acceptValues,
     assertViolationWithMessage = helper.assertViolationWithMessage;
 
@@ -30,7 +31,7 @@ buster.testCase('"none" condition', {
         acceptValues(condition, args, [false, null, [], {}, ['foo'], -7.5]);
     },
     'passes exceptions through': function () {
-        buster.referee.assert.exception(function () {
+        assert.exception(function () {
             condition(function () { x(); }).because('oh noes')('x');  // jshint ignore:line                
         }, 'ReferenceError');
     },