+++ test/condition/not-condition-test.js
@@ -5,7 +5,8 @@ var lewd = require('../../src/lewd'),
     errorMessages = require('../../src/messages'),
     helper = require('./../helper');
 
-var refuteValues = helper.refuteValues,
+var assert = buster.referee.assert,
+    refuteValues = helper.refuteValues,
     acceptValues = helper.acceptValues,
     assertViolationWithMessage = helper.assertViolationWithMessage;
 
@@ -25,7 +26,7 @@ buster.testCase('"not" condition', {
         acceptValues(condition, args, [0, -17.3, 'foo', 'X', false, [], {}, ['foo']]);
     },
     'passes exceptions through': function () {
-        buster.referee.assert.exception(function () {
+        assert.exception(function () {
             condition(function () { x(); })(null);  // jshint ignore:line                
         }, 'ReferenceError');
     },