+++ test/misc/coercion-test.js
@@ -6,33 +6,35 @@ var _ = require('lodash'),
 var lewd = require('../../src/lewd'),
     Condition = require('../../src/condition/Condition');
 
+var assert = buster.referee.assert;
+
 buster.testCase('coercion', {
     'coercion within objects': function () {
         var condition = lewd.object({ a: lewd(Boolean).coerce() }),
             obj = { a: 1 },
             result = condition(obj); 
 
-        buster.referee.assert.same(result, obj);
-        buster.referee.assert.equals(result, { a: true });
+        assert.same(result, obj);
+        assert.equals(result, { a: true });
     },
     'coercion within arrays': function () {
         var condition = lewd([lewd.integer().coerce()]),
             array = [42, 1.3, 0],
             result = condition(array);
 
-        buster.referee.assert.same(result, array);
-        buster.referee.assert.equals(result, [42, 1, 0]);
+        assert.same(result, array);
+        assert.equals(result, [42, 1, 0]);
     },
     '"integer condition"': {
         successful: function () {
-            buster.referee.assert.same(lewd.integer().coerce()(42), 42);
-            buster.referee.assert.same(lewd.integer().coerce()(42.3), 42);
+            assert.same(lewd.integer().coerce()(42), 42);
+            assert.same(lewd.integer().coerce()(42.3), 42);
         },
         failed: function () {
-            buster.referee.assert.exception(function () {
+            assert.exception(function () {
                 lewd.integer().coerce()('42');
             }, 'ConditionViolationException');
-            buster.referee.assert.exception(function () {
+            assert.exception(function () {
                 lewd.integer().coerce()('42x');
             }, 'ConditionViolationException');
         }
@@ -42,23 +44,23 @@ buster.testCase('coercion', {
             var date = new Date(),
                 value = lewd.isoDateTime().coerce()(date.toISOString());
             
-            buster.referee.assert(_.isDate(value), 'Value is a Date object');
-            buster.referee.assert.same(value.getTime(), date.getTime());
+            assert(_.isDate(value), 'Value is a Date object');
+            assert.same(value.getTime(), date.getTime());
         },
         failed: function () {
-            buster.referee.assert.exception(function () {
+            assert.exception(function () {
                 lewd.isoDateTime().coerce()('1234');
             }, 'ConditionViolationException');
         }
     },
     '"Boolean" condition': function () {
-        buster.referee.assert.same(lewd.Boolean().coerce()('1'), true);
-        buster.referee.assert.same(lewd.Boolean().coerce()(0), false);
-        buster.referee.assert.same(lewd.coerce(Boolean)('1'), true);
-        buster.referee.assert.same(lewd.coerce(Boolean)(0), false);
+        assert.same(lewd.Boolean().coerce()('1'), true);
+        assert.same(lewd.Boolean().coerce()(0), false);
+        assert.same(lewd.coerce(Boolean)('1'), true);
+        assert.same(lewd.coerce(Boolean)(0), false);
     },
     '"String" condition': function () {
         var value = lewd.String().coerce()(42);
-        buster.referee.assert.same(value, '42');
+        assert.same(value, '42');
     }
 });