+++ test/misc/path-test.js
@@ -3,7 +3,8 @@ var buster = require('buster');
 var lewd = require('../../src/lewd'),
     helper = require('../helper');
 
-var assertViolationAt = helper.assertViolationAt;
+var assert = buster.referee.assert,
+    assertViolationAt = helper.assertViolationAt;
 
 buster.testCase('paths in exceptions', {
     'top level': {
@@ -48,25 +49,67 @@ buster.testCase('paths in exceptions', {
             assertViolationAt(function () {
                 lewd.object({ a: { b: { c: null } } })({ a: { b: { c: 42 } } });
             }, ['a', 'b', 'c']); 
+        },
+        'invalid keys': function () {
+            assertViolationAt(function () {
+                lewd.array({
+                    $k: /a+/
+                })([{
+                    a: 0,
+                    aa: 1,
+                    b: 2
+                }]);
+            }, [0, '{b}']);
         }
     },
     'array': {
         '[Number]': function () {
             assertViolationAt(function () {
                 lewd.array(Number)([1, 2, '3', 4]);
-            }, ['#2']);
+            }, [2]);
         },
         '[[String]]': function () {
             assertViolationAt(function () {
                 lewd.array([String])([[], ['a'], [], ['3', 4, '5']]);
-            }, ['#3', '#1']);
+            }, [3, 1]);
         }
     },
     'combined': {
         '{ a: [{ b: [String] }]': function () {
             assertViolationAt(function () {
                 lewd.object({ a: [{ b: [String] }] })({ a: [{ b: ['x', 42] }] });
-            }, ['a', '#0', 'b', '#1']);
+            }, ['a', 0, 'b', 1]);
         }
+    },
+    'are valid JSON pointer references': function () {
+        var exception;
+        
+        try {
+            lewd.object({
+                'a/b': {
+                    1: [
+                        {
+                            $k: String,
+                            $v: {
+                                y: Number
+                            }
+                        }
+                    ]
+                }
+            })({
+                'a/b': {
+                    1: [{}, {
+                        'm~n': {
+                            y: null
+                        }
+                    }]
+                }
+            });
+        } catch (e) {
+            exception = e;
+        }
+
+        assert(exception);
+        assert.equals(exception.pathStr, '/a~1b/1/1/m~0n/y');
     }
 });